Completed day 10, part B
Part B was completed manually, based on the output from part A. The `day10-output.png` is a graphic rendering of the final map, where the starting point is marked in red, and the areas inside the pipe are filled with light green. The `day10-output.txt` file shows only the tiles inside the pipe, and the count of the "ground" character reveals the solution. This won't work for others with different puzzle data.